testing CI with new tests

+++ app/controllers/homepage_controller.rb
@@ -8,22 +8,6 @@ class HomepageController < ApplicationController
     respond_to do |format|
       format.html
       format.zip do
-        unless current_member
-          current_directory = Dir.pwd
-          zipname = "#{current_directory}/public/downloads/you_should_not_have_this.zip"
-          File.delete(zipname) if File.exist?(zipname)
-
-          ::Zip::File.open(zipname, ::Zip::File::CREATE) do |zipfile|
-            file = File.open('README.txt', 'w')
-            File.write('README.txt', 'SECURITY BREACH')
-            file.close
-            zipfile.add('README.txt', file)
-          end
-
-          send_file zipname
-          return
-        end
-
         @announcements = Announcement.all
         @events = Event.all
         @attendees = Attendee.all
@@ -50,10 +34,27 @@ class HomepageController < ApplicationController
           end
         end
 
+        # unless current_member
+        #   filenames.each do |filename|
+        #     File.delete(filename) if File.exist?(filename)
+        #   end
+
+        #   current_directory = Dir.pwd
+        #   zipname = "#{current_directory}/public/downloads/you_should_not_have_this.zip"
+        #   File.delete(zipname) if File.exist?(zipname)
+
+        #   ::Zip::File.open(zipname, ::Zip::File::CREATE) do |zipfile|
+        #     file = File.open('README.txt', 'w')
+        #     File.write('README.txt', 'SECURITY BREACH')
+        #     file.close
+        #     zipfile.add('README.txt', file)
+        #   end
+        # end
+
         send_file zipname
 
         filenames.each do |filename|
-          File.delete(filename)
+          File.delete(filename) if File.exist?(filename)
         end
       end
     end
